requirements.txt
atomicwrites==1.4.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coloredlogs==14.0
fpyutils==1.2.0
humanfriendly==8.2
idna==2.10
md-toc==7.0.3
pyreadline==2.1
PyYAML==5.3.1
readchar==2.0.1
requests==2.24.0
urllib3==1.25.10
